+++ packages/core/src/icons/components/RadioSelected.js
@@ -1,22 +1,25 @@
-import React from 'react';
+import * as React from 'react';
 
 export default function SvgRadioSelected(props) {
   return (
-    <svg
-      xmlns="http://www.w3.org/2000/svg"
-      viewBox="-1980 -5113 32 32"
-      width="1em"
-      height="1em"
-      {...props}
-    >
-      <path
-        data-name="\u9577\u65B9\u5F62 7901"
-        fill="transparent"
-        d="M-1980-5113h32v32h-32z"
+    <svg width="1em" height="1em" xmlns="http://www.w3.org/2000/svg" {...props}>
+      <rect
+        x={7}
+        y={7}
+        width={18}
+        height={18}
+        rx={9}
+        stroke="currentColor"
+        strokeWidth={2}
+        fill="none"
       />
-      <path
-        data-name="\u524D\u9762\u30AA\u30D6\u30B8\u30A7\u30AF\u30C8\u3067\u578B\u629C\u304D 8"
-        d="M-1958-5087h-12a4.005 4.005 0 01-4-4v-12a4 4 0 014-4h12a4 4 0 014 4v12a4 4 0 01-4 4zm-10.628-12.608l-2.122 2.122 5.656 5.657 8.486-8.485-2.122-2.122-6.364 6.364-3.535-3.536z"
+      <rect
+        x={11.5}
+        y={11.5}
+        width={9}
+        height={9}
+        rx={4.5}
+        stroke="currentColor"
       />
     </svg>
   );