+++ packages/form/src/utils/__tests__/parseSelectOptions.test.js
@@ -1,6 +1,7 @@
 import React from 'react';
 
 import SelectOption from 'src/SelectOption';
+import RadioSelectOption from 'src/RadioSelectOption';
 import parseSelectOptions from '../parseSelectOptions';
 
 it('reads options from React children of <SelectOption>s', () => {
@@ -14,6 +15,25 @@ it('reads options from React children of <SelectOption>s', () => {
     <SelectOption label="Foo" value="foo" checked />,
     <SelectOption label="Bar" value="bar" readOnly />,
     <SelectOption label="Meh" value="meh" />,
+  ];
+  expect(parseSelectOptions(multipleOptions)).toMatchObject([
+    { label: 'Foo', value: 'foo', checked: true },
+    { label: 'Bar', value: 'bar', readOnly: true },
+    { label: 'Meh', value: 'meh' },
+  ]);
+});
+
+it('reads options from React children of <RadioSelectOption>s', () => {
+  const singleOption = <RadioSelectOption label="Foo" value="foo" />;
+
+  expect(parseSelectOptions(singleOption)).toMatchObject([
+    { label: 'Foo', value: 'foo' },
+  ]);
+
+  const multipleOptions = [
+    <RadioSelectOption label="Foo" value="foo" checked />,
+    <RadioSelectOption label="Bar" value="bar" readOnly />,
+    <RadioSelectOption label="Meh" value="meh" />,
   ];
   expect(parseSelectOptions(multipleOptions)).toMatchObject([
     { label: 'Foo', value: 'foo', checked: true },